+++ boot/inject_commands.go
@@ -20,11 +20,9 @@ var commandSet = wire.NewSet(
 	cmd.NewHelloWorldCommand,
 	user.NewCreateCommand,
 	model.NewGeneratorCommand,
-	generators.NewControllerGeneratorCommand,
 	cmd.NewHttpServeCommand,
 	questapi.NewParseCommand,
 	cmd.NewRoutesListCommand,
-	generators.NewGenerateConfigurationCommand,
 	spire.NewMigrateCommand,
 	questapi.NewExampleTestCommand,
 	generators.NewRaceModelMapsCommand,
@@ -46,10 +44,8 @@ func ProvideCommands(
 	helloWorldCommand *cmd.HelloWorldCommand,
 	userCreateCommand *user.CreateCommand,
 	generateModelsCommand *model.GeneratorCommand,
-	generateControllersCommand *generators.ControllerGeneratorCmd,
 	httpServeCommand *cmd.HttpServeCommand,
 	routesListCommand *cmd.RoutesListCommand,
-	generateConfigurationCommand *generators.ConfigurationCommand,
 	spireMigrateCommand *spire.MigrateCommand,
 	questApiParseCommand *questapi.ParseCommand,
 	questExampleTestCommand *questapi.ExampleTestCommand,
@@ -69,10 +65,8 @@ func ProvideCommands(
 		helloWorldCommand.Command(),
 		userCreateCommand.Command(),
 		generateModelsCommand.Command(),
-		generateControllersCommand.Command(),
 		httpServeCommand.Command(),
 		routesListCommand.Command(),
-		generateConfigurationCommand.Command(),
 		spireMigrateCommand.Command(),
 		questApiParseCommand.Command(),
 		questExampleTestCommand.Command(),
