+++ boot/wire_gen.go
@@ -71,7 +71,6 @@ func InitializeApplication() (App, error) {
 	userUser := user.NewUser(resolver, encrypter, cache)
 	createCommand := user.NewCreateCommand(resolver, appLogger, encrypter, userUser)
 	generatorCommand := model.NewGeneratorCommand(db, appLogger)
-	controllerGeneratorCmd := generators.NewControllerGeneratorCommand(db, appLogger)
 	helloWorldController := controllers.NewHelloWorldController(db)
 	controller := auth.NewController(resolver, userUser, cache)
 	meController := user.NewMeController()
@@ -368,7 +367,6 @@ func InitializeApplication() (App, error) {
 	server := http.NewServer(appLogger, router, questHotReloadWatcher, launcher, clientManager)
 	httpServeCommand := cmd.NewHttpServeCommand(appLogger, server)
 	routesListCommand := cmd.NewRoutesListCommand(router)
-	configurationCommand := generators.NewGenerateConfigurationCommand(resolver, appLogger)
 	migrateCommand := spire.NewMigrateCommand(connections)
 	parseCommand := questapi.NewParseCommand(parseService)
 	exampleTestCommand := questapi.NewExampleTestCommand(examplesGithubSourcer)
@@ -383,7 +381,7 @@ func InitializeApplication() (App, error) {
 	launcherShimCmd := eqemuserver.NewLauncherShimCmd(launcher)
 	scrapeCommand := eqtraders.NewScrapeCommand(db, appLogger)
 	importCommand := eqtraders.NewImportCommand(db, appLogger)
-	v := ProvideCommands(helloWorldCommand, createCommand, generatorCommand, controllerGeneratorCmd, httpServeCommand, routesListCommand, configurationCommand, migrateCommand, parseCommand, exampleTestCommand, raceModelMapsCommand, changelogCommand, testFilesystemCommand, initCommand, changePasswordCommand, crashAnalyticsFingerprintBackfillCommand, updateCommand, launcherCmd, launcherShimCmd, scrapeCommand, importCommand)
+	v := ProvideCommands(helloWorldCommand, createCommand, generatorCommand, httpServeCommand, routesListCommand, migrateCommand, parseCommand, exampleTestCommand, raceModelMapsCommand, changelogCommand, testFilesystemCommand, initCommand, changePasswordCommand, crashAnalyticsFingerprintBackfillCommand, updateCommand, launcherCmd, launcherShimCmd, scrapeCommand, importCommand)
 	webBoot := desktop.NewWebBoot(appLogger, server, config)
 	bootApp := NewApplication(db, cache, v, resolver, connections, router, webBoot, init)
 	return bootApp, nil
