+++ internal/generators/controller_generator.go
@@ -1,3 +1,6 @@
+//go:build test
+// +build test
+
 package generators
 
 import (
