+++ internal/generators/sync_controllers_injector.go
@@ -1,3 +1,6 @@
+//go:build test
+// +build test
+
 package generators
 
 import (
