+++ internal/model/templates/crud_controller_types.go
@@ -1,0 +1,5 @@
+package crudcontrollers
+
+type BulkFetchByIdsGetRequest struct {
+	IDs []int `json:"ids"`
+}
