+++ internal/model/templates/inject_http_crud_controller.tmpl
@@ -1,0 +1,26 @@
+package boot
+
+import (
+	"github.com/Akkadius/spire/internal/http/crudcontrollers"
+	"github.com/Akkadius/spire/internal/http/routes"
+	"github.com/google/wire"
+)
+
+type crudControllers struct {
+	routes []routes.Controller
+}
+
+var httpCrudControllerSet = wire.NewSet(
+{{.NewControllers}}
+	provideCrudControllers,
+)
+
+func provideCrudControllers(
+{{.ControllersParam}}
+) *crudControllers {
+	return &crudControllers{
+		routes: []routes.Controller{
+{{.ControllersRegister}}
+		},
+	}
+}
