+++ internal/model/templates/vue_form.tmpl
@@ -1,0 +1,31 @@
+<template>
+  <vue-form-generator
+    v-if="model"
+    :schema="schema"
+    :model="model"
+    :options="formOptions"/>
+</template>
+
+<script>
+export default {
+  props: {
+    model: {
+      type: Object
+    }
+  },
+  data() {
+    return {
+      schema: {
+        fields:
+          {{.FormFields}}
+      },
+
+      formOptions: {
+        // validateAfterLoad: true,
+        // validateAfterChanged: true
+      }
+    };
+  }
+
+}
+</script>
